requirements.txt
flask==3.0.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
werkzeug==3.0.1
httpx==0.24.1
